📝 Deprecate `skip_download` in favour of `no_get`
`skip_download: true` allows us to prevent re-downloading in the
implicit get step that is run after a put step. However, this has the
added drawback of leaving us with an empty volume which means we need
to alias the resource for updating the pull-request.

We can instead use `no_get: true` on the put step to prevent the get
step from running entirely, removing the need to alias the resource as
the volume will be un-affected.

https://concourse-ci.org/put-step.html#schema.put.no_get